[NAYB-36] refactor : FindCouponResponse -> 내부 클래스 변경
