+++ src/main/java/com/prgrms/nabmart/domain/coupon/service/response/FindCouponsResponse.java
@@ -1,6 +1,7 @@
 package com.prgrms.nabmart.domain.coupon.service.response;
 
 import com.prgrms.nabmart.domain.coupon.Coupon;
+import java.time.LocalDate;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -11,5 +12,25 @@ public record FindCouponsResponse(List<FindCouponResponse> coupons) {
             .map(FindCouponResponse::from)
             .collect(Collectors.collectingAndThen(Collectors.toList(), FindCouponsResponse::new));
     }
-}
 
+    public record FindCouponResponse(
+        Long couponId,
+        String name,
+        String description,
+        Integer discount,
+        Integer minOrderPrice,
+        LocalDate endAt
+    ) {
+
+        public static FindCouponResponse from(final Coupon coupon) {
+            return new FindCouponResponse(
+                coupon.getCouponId(),
+                coupon.getName(),
+                coupon.getDescription(),
+                coupon.getDiscount(),
+                coupon.getMinOrderPrice(),
+                coupon.getEndAt()
+            );
+        }
+    }
+}
